requirements.txt
# Core dependencies
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
httpx>=0.24.1

# LLM and embeddings
openai>=1.3.0
anthropic>=0.5.0
transformers>=4.33.2
sentence-transformers>=2.2.2
torch>=2.0.1

# Vector storage
faiss-cpu>=1.7.4
chromadb>=0.4.13

# Data processing
numpy>=1.24.3
pandas>=2.1.0
scipy>=1.11.2
scikit-learn>=1.3.0

# Document handling
beautifulsoup4>=4.12.2
lxml>=4.9.3
pypdf>=3.16.1
pdfplumber>=0.10.2
markdown>=3.4.4

# Image processing
pillow>=10.0.1
opencv-python>=4.8.0.76
pydicom>=2.4.3

# Utilities
tqdm>=4.66.1
requests>=2.31.0
aiofiles>=23.2.1
tenacity>=8.2.3
rich>=13.5.2
pytest>=7.4.2 